+++ src/app.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UsersModule } from './users/users.module';
-import { User } from './users/entities/user.entity';
-import { Board } from './boards/entities/board.entity';
-import { Column } from './boards/entities/column.entity';
-import { Task } from './tasks/entities/task.entity';
+import { UsersModule } from './resources/users/users.module';
+import { User } from './resources/users/entities/user.entity';
+import { Board } from './resources/boards/entities/board.entity';
+import { Column } from './resources/boards/entities/column.entity';
+import { Task } from './resources/tasks/entities/task.entity';
 import {
   DB_HOST,
   DB_PORT,
@@ -14,8 +14,8 @@ import {
   DB_USERNAME,
   DB_PASSWORD,
 } from './common/config';
-import { BoardsModule } from './boards/boards.module';
-import { TasksModule } from './tasks/tasks.module';
+import { BoardsModule } from './resources/boards/boards.module';
+import { TasksModule } from './resources/tasks/tasks.module';
 
 @Module({
   imports: [
