+++ src/resources/boards/boards.service.ts
@@ -1,30 +1,46 @@
-import * as boardsRepo from './boards.memory.repository';
-import Board from './boards.model';
+import { Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { CreateBoardDto } from './dto/create-board.dto';
+import { UpdateBoardDto } from './dto/update-board.dto';
+import { Board } from './entities/board.entity';
 
-const getAll = async (): Promise<Board[]> => {
-  const boards = await boardsRepo.getAll();
+@Injectable()
+export class BoardsService {
+  constructor(
+    @InjectRepository(Board)
+    private boardsRepository: Repository<Board>
+  ) {}
 
-  return boards.map((board) => ({
-    ...board,
-    columns: board.columns.sort((a, b) => a.order - b.order),
-  }));
-};
+  create(createBoardDto: CreateBoardDto) {
+    return this.boardsRepository.save(createBoardDto);
+  }
 
-const getById = async (boardId: string): Promise<Board | undefined> => {
-  const board = await boardsRepo.getById(boardId);
+  async getAll() {
+    const boards = await this.boardsRepository.find();
 
-  if (!board) return undefined;
+    return boards.map((board) => ({
+      ...board,
+      columns: board.columns.sort((a, b) => a.order - b.order),
+    }));
+  }
 
-  return {
-    ...board,
-    columns: board.columns.sort((a, b) => a.order - b.order),
-  };
-};
+  async getById(id: string) {
+    const board = await this.boardsRepository.findOne(id);
 
-const create = (board: Board) => boardsRepo.create(Board.toDb(board));
+    if (!board) return undefined;
 
-const remove = (boardId: string) => boardsRepo.delete(boardId);
+    return {
+      ...board,
+      columns: board.columns.sort((a, b) => a.order - b.order),
+    };
+  }
 
-const update = (board: Board) => boardsRepo.update(Board.toDb(board));
+  update(id: string, updateBoardDto: UpdateBoardDto) {
+    return this.boardsRepository.save(updateBoardDto);
+  }
 
-export { getAll, create, remove as delete, getById, update };
+  remove(id: string) {
+    return this.boardsRepository.delete(id);
+  }
+}
