+++ src/resources/boards/entities/column.entity.ts
