--- src/resources/users/User.ts
@@ -1,26 +1,0 @@
-import { Entity, Column, PrimaryGeneratedColumn, OneToMany } from 'typeorm';
-
-import { Task } from '../tasks/Task';
-
-@Entity({ name: 'user' })
-class User {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column('varchar')
-  name = 'USER';
-
-  @Column('varchar')
-  login = 'user';
-
-  @Column('varchar')
-  password = 'P@55w0rd';
-
-  @OneToMany(() => Task, (task) => task.userId, {
-    eager: true,
-    cascade: true,
-  })
-  tasks?: Task[];
-}
-
-export { User };
